+++ core/src/labels/spriteLabel.cpp
@@ -7,7 +7,8 @@ SpriteLabel::SpriteLabel(Label::Transform _transform, glm::vec2 _size, LabelMesh
     int _vertexOffset, Label::Options _options, float _extrudeScale, glm::vec2 _anchor) :
     Label(_transform, _size, Label::Type::point, _mesh, {_vertexOffset, 4}, _options),
     m_extrudeScale(_extrudeScale), m_anchor(_anchor)
-{}
+{
+}
 
 void SpriteLabel::updateBBoxes(float _zoomFract) {
     glm::vec2 halfSize = m_dim * 0.5f;
