+++ core/src/labels/textLabel.cpp
@@ -2,23 +2,25 @@
 
 namespace Tangram {
 
+using namespace LabelProperty;
+
 TextLabel::TextLabel(Label::Transform _transform, Type _type, glm::vec2 _dim, TextBuffer& _mesh, Range _vertexRange,
-    Label::Options _options, FontContext::FontMetrics _metrics, int _nLines, Text::Anchor _anchor) :
+    Label::Options _options, FontContext::FontMetrics _metrics, int _nLines, Anchor _anchor) :
     Label(_transform, _dim, _type, static_cast<LabelMesh&>(_mesh), _vertexRange, _options),
     m_metrics(_metrics), m_nLines(_nLines)
 {
     if (m_type == Type::point) {
         glm::vec2 halfDim = m_dim * 0.5f;
         switch(_anchor) {
-            case Text::Anchor::left: m_anchor.x -= halfDim.x; break;
-            case Text::Anchor::right: m_anchor.x += halfDim.x; break;
-            case Text::Anchor::top: m_anchor.y -= halfDim.y; break;
-            case Text::Anchor::bottom: m_anchor.y += halfDim.y; break;
-            case Text::Anchor::bottom_left: m_anchor += glm::vec2(-halfDim.x, halfDim.y); break;
-            case Text::Anchor::bottom_right: m_anchor += halfDim; break;
-            case Text::Anchor::top_left: m_anchor -= halfDim; break;
-            case Text::Anchor::top_right: m_anchor += glm::vec2(halfDim.x, -halfDim.y); break;
-            case Text::Anchor::center: break;
+            case Anchor::left: m_anchor.x -= halfDim.x; break;
+            case Anchor::right: m_anchor.x += halfDim.x; break;
+            case Anchor::top: m_anchor.y -= halfDim.y; break;
+            case Anchor::bottom: m_anchor.y += halfDim.y; break;
+            case Anchor::bottom_left: m_anchor += glm::vec2(-halfDim.x, halfDim.y); break;
+            case Anchor::bottom_right: m_anchor += halfDim; break;
+            case Anchor::top_left: m_anchor -= halfDim; break;
+            case Anchor::top_right: m_anchor += glm::vec2(halfDim.x, -halfDim.y); break;
+            case Anchor::center: break;
         }
     }
 }
